refactor(view): extract design-tool source detection into claude_bundle_sources.cpp
claude_bundle.cpp had grown to 2,459 lines, bundling the runtime-import
harness together with the per-design-tool source-detection cluster. Any
work on a design-tool importer (v0.dev, Figma Make, Google Stitch, React
Native, Pencil) recompiled the entire harness.

Move the source-detection + runtime-JS-builder cluster into a new
sibling TU, core/view/src/claude_bundle_sources.cpp:

  * the shared v0_* foundation helpers (trim/lower/regex match,
    supported-surface gating, runtime-JS template builder);
  * the per-tool figma_*, stitch_*, rn_*, and pencil_* families;
  * the public parse_v0_dev_react / parse_figma_make_react /
    parse_stitch_react / parse_react_native_export / parse_pencil_react
    entry points (declared in design_import.hpp, so the public surface
    is unchanged).

The cross-TU interface is deliberately tiny: the public parse_* entry
points plus two helpers promoted to external linkage and declared in the
private design_import_internal.hpp — json_string_literal (the shared
JS-string escaper, still defined in claude_bundle.cpp) and
v0_html_attr_escape (the HTML-attribute escaper, defined in the new TU
and reused by parse_jsx_react). set_runtime_error and parse_jsx_react
stay in claude_bundle.cpp; they belong to the runtime harness, not the
source-detection cluster.

claude_bundle.cpp: 2,459 -> 1,434 lines. New TU: 1,088 lines.

Build (Debug, with Skia): pulp-view links clean. Tests all pass:
  pulp-test-design-import                   269 assertions / 30 cases
  pulp-test-design-import-react-runtime     183 assertions / 21 cases
  pulp-test-design-import-claude-runtime     11 assertions /  4 cases
  pulp-test-widget-bridge-runtime-import     36 assertions / 12 cases
