+++ core/view/CMakeLists.txt
@@ -249,6 +249,7 @@ target_sources(pulp-view PRIVATE
     src/design_codegen.cpp
     src/lock_to_source.cpp
     src/claude_bundle.cpp
+    src/claude_bundle_sources.cpp
     src/design_import_shortcuts.cpp
     src/design_import_designmd.cpp
     src/token_lock.cpp
