+++ tools/import-validation/source-contracts.json
@@ -138,7 +138,7 @@
       },
       "parser": {
         "file": "core/view/src/design_import.cpp",
-        "runtime_file": "core/view/src/claude_bundle.cpp",
+        "runtime_file": "core/view/src/claude_bundle_sources.cpp",
         "runtime": "parse_stitch_react",
         "static": "parse_stitch_html"
       },
@@ -320,7 +320,7 @@
       },
       "parser": {
         "file": "core/view/src/design_import.cpp",
-        "runtime_file": "core/view/src/claude_bundle.cpp",
+        "runtime_file": "core/view/src/claude_bundle_sources.cpp",
         "runtime": "parse_pencil_react",
         "static": "parse_pencil_json"
       },
@@ -419,7 +419,7 @@
       },
       "parser": {
         "file": "core/view/src/design_import.cpp",
-        "runtime_file": "core/view/src/claude_bundle.cpp",
+        "runtime_file": "core/view/src/claude_bundle_sources.cpp",
         "runtime": "parse_figma_make_react",
         "static": "parse_figma_json"
       },
@@ -511,7 +511,7 @@
       },
       "parser": {
         "file": "core/view/src/design_import.cpp",
-        "runtime_file": "core/view/src/claude_bundle.cpp",
+        "runtime_file": "core/view/src/claude_bundle_sources.cpp",
         "runtime": "parse_v0_dev_react",
         "static": "parse_v0_tsx"
       },
@@ -673,7 +673,7 @@
       },
       "parser": {
         "file": "core/view/src/design_import.cpp",
-        "runtime_file": "core/view/src/claude_bundle.cpp",
+        "runtime_file": "core/view/src/claude_bundle_sources.cpp",
         "runtime": "parse_react_native_export",
         "static": null
       },
